webkit debug-12 : This is the first self installable theme package  with Pods self loading and template default settings (still under test)

+++ inc/dependancy-check.php
@@ -23,12 +23,33 @@ if ( !defined( 'ABSPATH' ) ) exit;
 function cnrswebkit_dependancy_check(){
     global $messages;
     $messages = array(); 
+    
     // Check for pods ! 
-    if (! function_exists('pods')) {
+    if (! is_admin() && ! function_exists('pods')) {
+        /* We arrive here when theme has not been changed but an admin has unactivated Pods.
+         * This should never occurs !!
+         * This completely breaks frontend and backend websites; instead we issue a message.
+         * This message indicates that an admin must activate Pods.
+         */
         $messages[] = __( 'CNRS Web Kit requires Pods Framework/plugin. Please Ask an administrator to reactivate and/or install Pods .', 'cnrswebkit' );
-        if (is_admin() ){
-            $messages[] = __( 'It appears that Pods Framework has been uninstalled or unactivated. This breaks CNRS Webkit theme and causes frontend and backend freezing.', 'cnrswebkit' );  
-            $messages[] = '<a href="">' . __( 'Wordpress theme has been reset to Default theme.', 'cnrswebkit' ). '</a>';
+    }
+    
+    if (is_admin() ){
+        if (! function_exists('pods')) {
+            $messages[] = __( 'It appears that Pods Framework is not installed or unactivated. This breaks CNRS Webkit theme and causes frontend and backend freezing. For this reason, theme has been reset to Default Wordpress theme.', 'cnrswebkit' );  
+            /* Translators : "Pods – Custom Content Types and Fields" is "Pods - Types de contenus et champs personnalisés" in french */ 
+            $messages[] = '<a href="/wp-admin/plugin-install.php?tab=plugin-information&plugin=pods">' . __( 'Install "Pods – Custom Content Types and Fields".', 'cnrswebkit' ). '</a>';
+            $messages[] =  __( 'You must also install "Pods Migrates Package".', 'cnrswebkit' );
+            
+        } else {
+            // Check for Pods_Migrate_Packages
+            if ( ! class_exists( 'Pods_Migrate_Packages' ) ) {
+                $messages[] = __( 'It appears that Pods Migrates Package is not installed . This breaks CNRS Webkit theme and causes frontend and backend freezing. For this reason, theme has been reset to Default Wordpress theme.', 'cnrswebkit' );
+                $messages[] = '<a href="/wp-admin/admin.php?page=pods-components">' . __( 'Please enable  "Pods – Migrate: Packages".', 'cnrswebkit' ). '</a>';
+                
+            }}
+        if (! empty($messages)) {
+            // We must switch theme to default theme since this breaks frontend and backend websites
             switch_theme( WP_DEFAULT_THEME, WP_DEFAULT_THEME );
             unset( $_GET['activated'] );
         }
